fix(form-builder): restore focus/cursor after array change
this is an attempt at restoring focus when editing an non-keyed array item while another item is added/removed before.

since there's no way we can uniquely identify the array items, we're using the previous focus location to try to identify the item that the user was currently editing. After the array has updated with the new value, we look for the item of the nearest value, starting from the focused item and looking in either direction