wip(portable-text-editor): check value prop for equality against current editor before setting it
This might not be the final solution, but's it's an improvement for sure as proven by lowering timeout tresholds in the tests.

This is actually the problem with the IME problems we see too (typing in Korean).

The IME keyboard doesn't trig the throttled state (no activity for 500 ms) and a new prop is set and confuses the IME apparently.